# Conflicts:
#   src/kernel/process/schedualer.c

+++ src/kernel/process/schedualer.c
@@ -15,7 +15,7 @@ static uint32_t kernel_cr3; // physical address of kernel page directory
 
 void scheduler_add(PCB *p) {
   p->pid = next_pid++;
-  p->cr3 = kernel_cr3;
+  p->cr3 = (uint32_t)kernel_cr3;
   tasks[task_count++] = p;
 }
